mnist_torch_vanilla/requirements.txt
metaflow-checkpoint
torchvision
torch